web_service/requirements.txt
flask
pandas
numpy
openpyxl
python-dotenv
Topsis-Vikas-102303451
